# docs/rapidapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Quantura",
    "version": "1.0.0",
    "description": "Probabilistic market forecasts, verified market discovery, observed historical data, and screener quantiles. Forecasts run asynchronously: create a job, poll its status, then download the final ensemble. Enterprise agreements define compute admission and API quotas. Market data remains subject to provider coverage and licensing. Research estimates are not guaranteed returns.",
    "contact": {
      "name": "Quantura",
      "url": "https://quantura.studio/contact"
    },
    "termsOfService": "https://quantura.studio/terms"
  },
  "servers": [
    {
      "url": "https://quantura.studio/api",
      "description": "Quantura production origin behind RapidAPI"
    }
  ],
  "security": [
    {
      "rapidApiKey": []
    }
  ],
...
  "paths": {
...
    "/v1/market-data/dukascopy/instruments": {
      "servers": [
        {
          "url": "https://quantura.studio/api"
        }
      ],
      "get": {
        "tags": [
          "Q Search"
        ],
        "security": [
          {
            "rapidApiKey": []
          }
        ],
        "operationId": "listDukascopyInstruments",
        "summary": "Browse every instrument in the complete Dukascopy catalog",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response (illustrative example)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "ok": true,
                  "provider": "dukascopy",
                  "instruments": [],
                  "total": 1504,
                  "next_offset": 100,
                  "stale": false
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "ok": false,
                  "error": "event_query_invalid",
                  "message": "Choose a provider event and a valid continuation cursor."
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/market-data/stocks/history": {
      "servers": [
        {
          "url": "https://quantura.studio/api"
        }
      ],
      "post": {
        "tags": [
          "Q Download"
        ],
        "security": [
          {
            "rapidApiKey": []
          }
        ],
        "operationId": "downloadQuanturaStockHistory",
        "summary": "Download split-adjusted stock history with automatic provider routing",
        "description": "Q Download previews JSON and exports that exact snapshot as CSV/JSON. Provider coverage varies; missing bars are not fabricated. format=csv returns a CSV attachment. No uploaded workspace files are exposed here.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "symbol"
                ],
                "properties": {
                  "symbol": {
                    "type": "string"
                  },
                  "contractSymbol": {
                    "type": "string"
                  },
                  "source": {
                    "type": "string",
                    "enum": [
                      "auto",
                      "alpaca",
                      "dukascopy"
                    ],
                    "default": "auto"
                  },
                  "price_side": {
                    "type": "string",
                    "enum": [
                      "bid",
                      "ask"
                    ],
                    "default": "bid"
                  },
                  "page_mode": {
                    "type": "boolean",
                    "description": "Dukascopy: retrieve bounded candle file pages; follow next_cursor using the same settings."
                  },
                  "cursor": {
                    "type": "string"
                  },
                  "start": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "end": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "timeframe": {
                    "type": "string",
                    "enum": [
                      "1min",
                      "5min",
                      "15min",
                      "30min",
                      "1h",
                      "4h",
                      "1D",
                      "1W-MON",
                      "1MS",
                      "1Min",
                      "5Min",
                      "15Min",
                      "30Min",
                      "1Hour",
                      "4Hour",
                      "1Day",
                      "1Week",
                      "1Month",
                      "1m",
                      "5m",
                      "15m",
                      "30m",
                      "1w"
                    ]
                  },
                  "limit": {
                    "type": "integer",
                    "enum": [
                      500,
                      1000,
                      1500,
                      2000,
                      50000
                    ]
                  },
                  "adjustment": {
                    "type": "string",
                    "enum": [
                      "raw",
                      "split",
                      "all"
                    ]
                  },
                  "session": {
                    "type": "string",
                    "enum": [
                      "regular",
                      "extended"
                    ]
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "json",
                      "csv"
                    ]
                  }
                }
              },
              "example": {
                "symbol": "AAPL",
                "source": "auto",
                "timeframe": "1Day",
                "limit": 500,
                "adjustment": "split",
                "format": "json"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON rows or CSV attachment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "ok": true,
                  "provider": "alpaca",
                  "sourceRequested": "auto",
                  "fallbackUsed": false,
                  "symbol": "AAPL",
                  "timeframe": "1Day",
                  "count": 1,
                  "rows": [
                    {
                      "timestamp": "2026-09-18T04:00:00Z",
                      "open": 100,
                      "high": 102,
                      "low": 99,
                      "close": 101,
                      "volume": 1000
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "ok": false,
                  "error": "event_query_invalid",
                  "message": "Choose a provider event and a valid continuation cursor."
                }
              }
            }
          },
          "422": {
            "description": "Invalid request or provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "ok": false,
                  "error": "event_query_invalid",
                  "message": "Choose a provider event and a valid continuation cursor."
                }
              }
            }
          },
          "429": {
            "description": "Invalid request or provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "ok": false,
                  "error": "event_query_invalid",
                  "message": "Choose a provider event and a valid continuation cursor."
                }
              }
            }
          },
          "503": {
            "description": "Invalid request or provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "ok": false,
                  "error": "event_query_invalid",
                  "message": "Choose a provider event and a valid continuation cursor."
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "rapidApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RapidAPI-Key",
        "description": "Your RapidAPI application key. RapidAPI supplies the API-specific gateway credential privately. A private enterprise CUSTOM subscription is required."
      }
    },
...
  }
}